msm: SSR: Merge start/stop ops with powerup/shutdown ops
The start/stop ops were introduced to provide a clear migration
path for merging SSR drivers and PIL drivers together. Now that
the merge is over we can drop the start/stop ops and use the
powerup/shutdown ops in their place. Mostly this means we disable
the irq during probe so that powerup() can enable it and then
drop the start/stop callbacks.

Change-Id: I787b60a93a31a4085c4118292ea8a163df175765
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>

Conflicts:
	arch/arm/mach-msm/pil-pronto.c
	arch/arm/mach-msm/pil-q6v5-lpass.c
	arch/arm/mach-msm/pil-q6v5-mss.c
	arch/arm/mach-msm/pil-riva.c
	arch/arm/mach-msm/subsystem_restart.c
	sound/soc/codecs/msm8x10-wcd.c