hwmon: epm_adc: Fix clear buffer command
Currently when clients request to clear the PSOC buffer
the result returned to the user is not valid. This is
because the response byte that indicates the status
of the PSOC is the second byte and not the third.
Fix it by reading the appropriate byte to indicate
the status of the command request.

Change-Id: I1604274619bb881ecc2e0afc46979ae4eb85a154
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>