ASoC: wcd9xxx: fix HPHL DAC path enable sequence
The dac_ref_en bit should not be turned ON by codec
driver when headset is inserted. Setting the
dac_ref_en bit, and not resetting it when headset is
removed, will cause the DAPM framework to keep HPHL
DAC path connected.
This change returns the ownership of setting the
dac_ref_en bit to DAPM framework and fixes the HPHL DAC
path enablement issue when headset is inserted.

CRs-fixed: 483600
Change-Id: Iabfe09a82507579a443b87458dc8dba0d2cd5ce5
Signed-off-by: Banajit Goswami <bgoswa@codeaurora.org>